requirements.txt
# App requirements
flask-socketio>=2.9.2
flask-restplus>=0.10.1
flask-cors>=3.0.3
eventlet>=0.21.0
hdfs3>=0.2.0
starbase>=0.3.3
happybase>=1.1.0
pydicom>=0.9.9
numpy>=1.13.1
celery>=4.1.0
Pillow>=4.3.0
retrying>=1.3.3
flask-user==0.6.13
psycopg2

# Dev requirements - these should always be latest
mypy
pylint
flake8
pytest
pytest-cov
pytest-mock
jupyter
matplotlib
